Ensure shapeDistTraveled is set (it is used in PresentationServiceImpl.include) and it is a double

+++ onebusaway-transit-data-federation/src/main/java/org/onebusaway/transit_data_federation/services/tripplanner/TripPlannerGraph.java
@@ -72,6 +72,6 @@ public interface TripPlannerGraph {
   public boolean updateStopTime(AgencyAndId tripId, AgencyAndId stopId, int originalArrivalTime, int originalDepartureTime,
                                 int newArrivalTime, int newDepartureTime);
 
-  public boolean insertStopTime(AgencyAndId tripId, AgencyAndId stopId, int arrivalTime, int departureTime, int shapeDistanceTravelled);
+  public boolean insertStopTime(AgencyAndId tripId, AgencyAndId stopId, int arrivalTime, int departureTime, double shapeDistanceTravelled);
 
 }
